ferbos_addon_query/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
flask-socketio==5.3.6
flask-cors==4.0.0
python-socketio==5.9.0
eventlet==0.33.3
requests==2.31.0
